+++ Sources/MapboxSearchUI/MapboxSearchController.swift
@@ -156,9 +156,14 @@ public class MapboxSearchController: UIViewController {
     public required init(accessToken: String, configuration: Configuration = Configuration()) {
         self.categorySearchEngine = CategorySearchEngine(
             accessToken: accessToken,
-            locationProvider: configuration.locationProvider
+            locationProvider: configuration.locationProvider,
+            apiType: .searchBox
         )
-        self.searchEngine = SearchEngine(accessToken: accessToken, locationProvider: configuration.locationProvider)
+        self.searchEngine = SearchEngine(
+            accessToken: accessToken,
+            locationProvider: configuration.locationProvider,
+            apiType: .searchBox
+        )
         self.configuration = configuration
 
         super.init(nibName: nil, bundle: .mapboxSearchUI)
@@ -173,8 +178,14 @@ public class MapboxSearchController: UIViewController {
     /// - Parameters:
     ///   - configuration: configuration for search and categorySearch engines.
     public required init(configuration: Configuration = Configuration()) {
-        self.categorySearchEngine = CategorySearchEngine(locationProvider: configuration.locationProvider)
-        self.searchEngine = SearchEngine(locationProvider: configuration.locationProvider)
+        self.categorySearchEngine = CategorySearchEngine(
+            locationProvider: configuration.locationProvider,
+            apiType: .searchBox
+        )
+        self.searchEngine = SearchEngine(
+            locationProvider: configuration.locationProvider,
+            apiType: .searchBox
+        )
         self.configuration = configuration
 
         super.init(nibName: nil, bundle: .mapboxSearchUI)