+++ Tests/MapboxSearchUITests/Integration/FavoritesIntegrationTestCase.swift
@@ -1,6 +1,7 @@
 import XCTest
 
-class FavoritesIntegrationTestCase: MockServerUITestCase {
+// Rename to UITestCase
+class FavoritesIntegrationTestCase: MockSearchBoxUITestCase {
     func testAddRemoveFavorite() throws {
         try server.setResponse(.suggestMinsk)
         try server.setResponse(.retrieveMinsk)